feat: change js script to stimulus
